requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
joblib==1.3.2
python-dotenv==1.0.0